issue 300: fix job_id type in the r2 merge handler (test build)

+++ crates/app/src/supervisor.rs
@@ -2726,7 +2726,7 @@ impl Supervisor {
                         dry_run,
                         max_groups: *max_groups,
                         expect_groups,
-                        job_id: Some(job_id),
+                        job_id: Some(job_id as i64),
                         stop: &stop,
                     })
                     .await
